requirements.txt
black==19.10b0
eth-brownie>=1.6.2,<2.0.0
flake8==3.7.9
isort==4.3.21
tox==3.14.3
